Change `epoch` to use UTC instead of local timezone
Prior to this change, the tests fail if running in a timezone west of UTC.
This change makes a functional difference to the `epoch` extension:
it now interprets epoch values as being UTC, where previously they were
interpreted in the local timezone.

Tests are unchanged but now pass regardless of the timezone they run in.